Adding zap
Zap is a solid logger, has levels, and should be configurable to route wherever logs need to go. Not all log statements ported yet.
