+++ api/Makefile
@@ -25,4 +25,4 @@ e2e:
 	go test -cpuprofile cpu.prof -memprofile mem.prof -bench -v ./e2e
 
 test:
-	go test -v ./...
+	go test -test.fullpath=true -timeout 300s -run ^TestE2E$ github.com/navapbc/its-log/e2e
